elevenlabs-customer-service-agent/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
asyncpg>=0.29.0

# Redis
redis>=5.0.0

# HTTP client (for tools that call out)
httpx>=0.26.0

# Config & validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-frontmatter>=1.0.0

# Logging
structlog>=24.0.0

# Dev / test
pytest>=7.0.0
pytest-asyncio>=0.23.0

langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
langchain_huggingface
sentence-transformers
psycopg[binary,pool]>=3.2.0

# Email
sendgrid>=6.11.0

#Milvus
protobuf
grpcio-tools
pymilvus>=2.5.0

# RabbitMQ
pika
aio-pika